Change HTTP request closure to displayError when status code is not 200 instead of when !response.result.isSuccess because this was not triggering for 500 or 400 status codes which should result in an error displayed
