+++ BenchwarmersApp/SupportGroupPicker.swift
@@ -22,7 +22,7 @@ class SupportGroupPicker: BenchwarmersPicker  {
     func getSupportGroups(controller: UIViewController, completion: @escaping () -> Void) {
         Alamofire.request(supportGroupsUrl, method: .get).responseJSON {
             response in
-            if response.result.isSuccess {
+            if response.response?.statusCode == 200 {
                 print("received support groups data")
                 self.supportGroups = JSON(response.result.value!)["data"].arrayValue
                 self.selection = self.supportGroups![0]
